refactor: move event types to shared folder
