+++ src/back/models/Event.ts
@@ -1,46 +1,7 @@
 import { Model } from 'decentraland-gatsby/dist/entities/Database/model'
 import { SQL, table } from 'decentraland-gatsby/dist/entities/Database/utils'
 
-export type CommonEventAttributes = {
-  id: string
-  address: string
-  created_at: Date
-}
-
-export enum EventType {
-  Voted = 'voted',
-  ProposalCreated = 'proposal_created',
-  UpdateCreated = 'update_created',
-  Commented = 'commented',
-}
-
-export type Event = VotedEvent | ProposalCreatedEvent | UpdateCreatedEvent | CommentedEvent
-
-export type VotedEvent = {
-  event_type: EventType.Voted
-  event_data: VoteEventData
-} & CommonEventAttributes
-
-export type ProposalCreatedEvent = {
-  event_type: EventType.ProposalCreated
-  event_data: ProposalEventData
-} & CommonEventAttributes
-
-export type UpdateCreatedEvent = {
-  event_type: EventType.UpdateCreated
-  event_data: UpdateCreatedEventData
-} & CommonEventAttributes
-
-export type CommentedEvent = {
-  event_type: EventType.Commented
-  event_data: ProposalEventData
-} & CommonEventAttributes
-
-type VoteEventData = { choice: string } & ProposalEventData
-type ProposalEventData = { proposal_id: string; proposal_title: string }
-type UpdateCreatedEventData = {
-  update_id: string
-} & ProposalEventData
+import { Event } from '../../shared/types/events'
 
 export default class EventModel extends Model<Event> {
   static tableName = 'events'
